requirements.txt
# Universal API Wrapper
litellm

# Memory
mem0ai
chromadb

# Channels
python-telegram-bot
fastapi
uvicorn
python-multipart

# Utilities
requests
python-dotenv
duckduckgo-search
PyYAML
